feat: convert select emoji to functional component
https://youtrack.weseek.co.jp/issue/GW-7652
- Rename EmojiPicker to Emoji
- Implement functional component to select emoji
- Change emoji search pattern
